requirements.txt
python-telegram-bot==20.7
fastapi==0.95.2
uvicorn==0.22.0
